add-to-cart multiple items, login not required
You can add multiple items to the cart without logging in. The values are being stored in cookies as cart_items.

As soon as you login, the items you added before login, gets stored into your database. The cookie gets deleted.

After logging in, the new items that you add to the cart will be stored in the database only.