view cart items
The customer can view the items they have added to their cart with or without logging in. Logged in users cart items are displayed from the database, however, for the users who have yet not logged in, the cart items are displayed from cookies.

If the user logs in, the cookies' cart_items will get stored into the database.