/place-order/ draft without payment method
If the customer has no address with type primary, they will be redirected to /customer/add-details/ else the order will be placed.

A completely unique order number is generated based on the ordering time, and a 10 digit random number.

The ordered items' total price is calculated and stored, the stock is updated.